refactor(evalhub): move EvalHubV1/V1Alpha1 classes to shared utils
